cobain

+++ application/controllers/UnitProperti.php
@@ -69,13 +69,21 @@ class UnitProperti extends CI_Controller {
         $data['img'] = getCompanyLogo();
         $this->pages("unit_properti/view_tambah_unit",$data); 
     }
+    public function multiTambah() //Menampilkan Form Tambah
+    {
+        $data['title'] = 'Tambah Unit';
+        $data['menus'] = $this->rolemenu->getMenus();
+        $data['js'] = $this->rolemenu->getJavascript(6); //Jangan DIUbah !!
+        $data['img'] = getCompanyLogo();
+        $this->pages("unit_properti/view_tambah_multiunit",$data); 
+    }
     public function detail_unit($id) //Menampilkan Form Tambah
     {
         $data['title'] = 'Detail Unit';
         $data['menus'] = $this->rolemenu->getMenus();
         $data['js'] = $this->rolemenu->getJavascript(6); //Jangan DIUbah !!
         $data['img'] = getCompanyLogo();
-        $data['unit'] = $this->Munit->getUnitWithId($id);
+        $data['unit'] = $this->Munit->getData("*","tbl_unit_properti",["id_unit"=>$id])->row();
         $this->pages("unit_properti/view_detail_unit",$data); 
     }
 
@@ -88,11 +96,11 @@ class UnitProperti extends CI_Controller {
         $this->validate();
         if ($this->form_validation->run() == false) {
             foreach ($_POST as $key => $value) {
-                $data['success'] = false;
                 $data['msg'][$key] = form_error($key);
             }
         }
         else{
+            $inputData = $this->inputData();
             $id_properti = $this->session->userdata('id_properti');
             $jml_unit = $this->Munit->getJumlahUnit($id_properti); 
             $jml_properti = $this->Munit->getJumlahProperti($id_properti);
@@ -106,34 +114,30 @@ class UnitProperti extends CI_Controller {
                 $data['text'] = "Unit Properti sudah mencapai batas yang ditentukan";
                 $data['type'] = "error";
             }else{
-                $config['upload_path'] = './assets/uploads/images/unit_properti';
-                $config['allowed_types'] = 'gif|jpg|png';
-                $config['encrypt_name'] = true;
-                $config['max_size']  = '2048';
-                $config['max_width']  = '1024';
-                $config['max_height']  = '768';
+                $config = $this->imageInit('./assets/uploads/images/unit_properti/');
                 $this->load->library('upload', $config);
                 if ($_FILES['foto']['name'] != "") {
                     if ($this->upload->do_upload('foto')){
                         $img = $this->upload->data();
-                        $input = $this->input( $img['file_name']);
-                        $this->Munit->insertUnit($input);
-                        $data['success'] = true;
-                        $data['title'] = "Berhasil";
-                        $data['text'] = "Unit Properti berhasil ditambah";
-                        $data['type'] = "success";
+                        $inputData += ["foto_unit"=>$img['file_name']];
+                        $query = $this->Munit->insertData($inputData,"unit_properti");
+                        if ($query) {
+                            $data['success'] = true;
+                            $data['title'] = "Berhasil";
+                            $data['text'] = "Unit Properti berhasil ditambah";
+                            $data['type'] = "success";
+                        }
                     }
                     else {
                         $data['error'] = $this->upload->display_errors();
                         $data['title'] = "Gagal";
                         $data['text'] = "Unit Properti berhasil ditambah";
                         $data['type'] = "error";
-                        $data['success'] = false;
                     }
                 }
                 else{
-                    $input = $this->input("");
-                    $act = $this->Munit->insertUnit($input);
+                    $inputData += ["foto_unit"=>"default.jpg"];
+                    $act = $this->Munit->insertData($inputData,"unit_properti");
                     if ($act) {
                         $data['success'] = true;
                         $data['title'] = "Berhasil ";
@@ -141,12 +145,69 @@ class UnitProperti extends CI_Controller {
                         $data['type'] = "success";
                     }
                     else{
-                        $data['success'] = false;
                         $data['title'] = "Gagal";
                         $data['text'] = "Gagal Menambahkan";
                         $data['type'] = "error";
                     }
                 }
+            }
+        }
+        return $this->output->set_output(json_encode($data));
+    }
+    public function coreMultiTambah() //Unit Core Tambah
+    {
+        $data = [
+            "success" => false,
+            'msg' => [],
+        ];
+        $this->validate();
+        $this->form_validation->set_rules('txt_blok_awal','Blok Awal','trim|required|numeric|max_length[11]');
+        $this->form_validation->set_rules('txt_jumlah_blok','Blok Akhir','trim|required|numeric|max_length[11]');
+        if ($this->form_validation->run() == false) {
+            foreach ($_POST as $key => $value) {
+                $data['msg'][$key] = form_error($key);
+            }
+        }
+        else{
+            $awal = (int) $this->input->post("txt_blok_awal",true);
+            $jumlah = (int) $this->input->post("txt_jumlah_blok",true);
+            $id_properti = $this->session->userdata("id_properti");
+            $query = $this->Munit->getData("jumlah_unit","tbl_properti",["id_properti"=>$id_properti])->row();
+            $query_unit = $this->Munit->getData("COUNT(id_unit) as jumlah_unit","tbl_unit_properti",["id_properti"=>$id_properti])->row();
+            if ($jumlah > 0) {
+                if ($query_unit->jumlah_unit < $query->jumlah_unit) {
+                    $sisa = ($query->jumlah_unit - $query_unit->jumlah_unit);
+                    if ($jumlah > $sisa) {
+                        $data["error"] = "Jumlah Terlalu Banyak";
+                    }else{
+                        $nama = $this->input->post("txt_nama",true);
+                        $input = [
+                            "nama_unit"=> $this->input->post("txt_nama",true),
+                            "id_properti"=> $id_properti,
+                            'type'=>$this->input->post('txt_type',true),
+                            'luas_tanah'=>$this->input->post('txt_tanah',true),
+                            'luas_bangunan'=>$this->input->post('txt_bangunan',true),
+                            'harga_unit'=>$this->input->post('txt_harga',true),
+                            'alamat_unit'=>$this->input->post('txt_alamat',true),
+                            'tgl_insert'=>date("Y-m-d"),
+                            'status_unit'=>"belum terjual",
+                            'deskripsi'=>$this->input->post('txt_desc',true),
+                            'foto_unit' => "default.jpg",
+                            'id_user' => $this->session->userdata("id_user")
+                        ];
+                        $total = $awal + $jumlah;
+                        for($i = $awal ; $i < $total ; $i++) {
+                            $input["nama_unit"] = $nama.$i;
+                            $this->Munit->insertData($input,"unit_properti");
+                        }
+                        $data["success"] = true;
+                        $data["url"] = base_url("unitproperti");
+                    }
+                }else{
+                    $data["error"] = "Unit Sudah FUll";
+                }
+            }else{
+                $data["error"] = "Jumlah harus lebih dari 0";
             }
         }
         return $this->output->set_output(json_encode($data));
@@ -166,43 +227,34 @@ class UnitProperti extends CI_Controller {
             }
         }
         else{
-            $config['upload_path'] = './assets/uploads/images/unit_properti';
-            $config['allowed_types'] = 'gif|jpg|png';
-            $config['encrypt_name'] = true;
-            $config['max_size']  = '2048';
-            $config['max_width']  = '1024';
-            $config['max_height']  = '768';
-            $data['id'] = $this->input("");
+            $id = $this->input->post("txt_id",true);
+            $inputData = $this->inputData();
+            $config = $this->imageInit("./assets/uploads/images/unit_properti/");
             $this->load->library('upload', $config);
             if ($_FILES['foto']['name'] != "") {
                 if ($this->upload->do_upload('foto')){
-                    $link = $this->Munit->getImage($this->input->post('txt_id'));
-                    $path = "./assets/uploads/images/unit_properti/".$link->foto_unit;
+                    $link = $this->Munit->getData("foto_unit","unit_properti",["id_unit"=>$id])->row();
+                    $path = base_url("assets/uploads/images/unit_properti/".$link->foto_unit);
                     if (file_exists($path)) {
-                        unlink($path);
-                    }
-                    else{
-                        $data['error'] = "tidak ditemukan";
+                        if ($link != "default.jpg") {
+                            unlink($path);
+                        }
                     }
                     $img = $this->upload->data();
-                    $input = $this->input($img['file_name'],$this->input->post('txt_id')); 
-                    $this->Munit->updateUnit($input);
-                    $data['success'] = true;
+                    $inputData += ["foto_unit"=>$img["file_name"]];
+                    $query = $this->Munit->updateData($inputData,"unit_properti",["id_unit"=>$id]);
+                    if ($query) {
+                        $data['success'] = true;
+                    }
                 }
                 else {
                     $data['error'] = $this->upload->display_errors();
-                    $data['success'] = false;
                 }
             }
-            else{
-                $input = $this->input("",$this->input->post('txt_id')); 
-                $act = $this->Munit->updateUnit($input);
+            else{ 
+                $act = $this->Munit->updateData($inputData,"unit_properti",["id_unit"=>$id]);
                 if ($act) {
                     $data['success'] = true;
-                    $data['act'] = $act;
-                }
-                else{
-                    $data['success'] = false;
                     $data['act'] = $act;
                 }
             }
@@ -215,24 +267,44 @@ class UnitProperti extends CI_Controller {
         $data = [
             "success" => false,
         ];
-        $input = $this->input->post('id_unit');
-        $query = $this->Munit->hapusData($input);
-        if ($query) {
-            $link = $this->Munit->getImage($input);
-            if ($link ['foto_unit'] != "") 
+        $input = $this->input->post('id_unit',true);
+        $image = $this->Munit->getData("foto_unit","unit_properti",["id_unit"=>$input])->row_array();
+        if ($image["foto_unit"] != null) {
+            if ($image['foto_unit'] != "default.jpg") 
             {
-                $rows = $link['foto_unit'];
-            }else{
-                $rows ="as";
+                $path = "./assets/uploads/images/unit_properti/".$image["foto_unit"];
+                if (file_exists($path)) {
+                    unlink($path);
+                }
             }
-            $path = "./assets/uploads/images/unit_properti/".$rows;
-            if (file_exists($path)) {
-                unlink($path);
-            }
-        }else{
-            $data['success'] = false;
+        }
+        $query = $this->Munit->deleteData("unit_properti",["id_unit"=>$input]);
+        if ($query) {
+            $data["success"] = true;
         }
         $this->output->set_output(json_encode($data));
+    }
+    public function getJumlahData()
+    {
+        $data = ["success"=>false,"jumlah"=>"bf"]; // f1 ="Unit Sudah Full" AND f2 = "Jumlah Unit Yang mau dibuat kebanyakan" AND bf = "BELUM FULL";
+        $jumlah = (int) $this->input->post("jumlah",true);
+        $id = $this->session->userdata("id_properti");
+        if ($jumlah != null) {
+            $query = $this->Munit->getData("jumlah_unit","tbl_properti",["id_properti"=>$id])->row();
+            $query_unit = $this->Munit->getData("COUNT(id_unit) as jumlah_unit","tbl_unit_properti",["id_properti"=>$id])->row();
+            if ($query_unit->jumlah_unit < $query->jumlah_unit) {
+                $sisa = ($query->jumlah_unit - $query_unit->jumlah_unit);
+                if ($jumlah > $sisa) {
+                    $data["success"] = true;
+                    $data["jumlah"] = "f2";
+                }
+            }else{
+                $data["success"] = true;
+                $data["jumlah"] = "f1";
+            }
+        }
+        return $this->output->set_content_type('application/json')->set_output(json_encode($data));
+        
     }
     // This function is private. so , anyone cannot to access this function from web based
     private function pages($core_page,$data){
@@ -254,19 +326,32 @@ class UnitProperti extends CI_Controller {
         $this->form_validation->set_rules('txt_desc','Deskripsi','trim|required');
         $this->form_validation->set_error_delimiters('<div class="invalid-feedback">','</div>');
     }
-    private function input($img)
+    private function inputData()
     {
-        return [
-            'id'=>$this->input->post('txt_id',true),
-            'nama'=>$this->input->post('txt_nama',true),
+        $input = [
+            "nama_unit"=> $this->input->post("txt_nama",true),
+            "id_properti"=> $this->session->userdata("id_properti"),
             'type'=>$this->input->post('txt_type',true),
-            'tanah'=>$this->input->post('txt_tanah',true),
-            'bangunan'=>$this->input->post('txt_bangunan',true),
-            'harga'=>$this->input->post('txt_harga',true),
-            'alamat'=>$this->input->post('txt_alamat',true),
+            'luas_tanah'=>$this->input->post('txt_tanah',true),
+            'luas_bangunan'=>$this->input->post('txt_bangunan',true),
+            'harga_unit'=>$this->input->post('txt_harga',true),
+            'alamat_unit'=>$this->input->post('txt_alamat',true),
+            'tgl_insert'=>date("Y-m-d"),
+            'status_unit'=>"belum terjual",
             'deskripsi'=>$this->input->post('txt_desc',true),
-            'img' => $img
+            'id_user' => $this->session->userdata("id_user")
         ];
+        return $input;
+    }
+    private function imageInit($path)
+    {
+        $config['upload_path'] = $path;
+        $config['allowed_types'] = 'gif|jpg|png';
+        $config['encrypt_name'] = true;
+        $config['max_size']  = '2048';
+        $config['max_width']  = '1024';
+        $config['max_height']  = '768';
+        return $config;
     }
 }
